refactor(osedges): split useOsEdges into per-edge subscribers
Split the 619-LOC useOsEdges.ts into a thin facade plus a directory of
narrow subscribers under src/hooks/osEdges/, one per OS edge:

  shellStreams    PTY output / exit / title
  agentReload     supervisor agent-reloaded handshake re-prime
  agentCrash      supervisor agent-crashed + auto-restart
  agentStderr     filtered stderr -> chat system message
  menu            native menu activations + ext:* routing
  dragDrop        OS file drop -> PTY or agent draft
  clipboardPaste  image paste -> save_paste_image + @path token

Each subscriber takes a narrow Deps slice (only the refs and callbacks
it actually uses) and returns a cleanup function. The facade composes
them inside the single boot useEffect and tears down on unmount.

No behavior change. App.tsx is untouched: useOsEdges keeps its
import path and signature.
